+++ frontend_new/app/static/js/config.js
@@ -1,0 +1,3 @@
+window.ATLAS_CFG = {
+  orchestratorBase: 'http://127.0.0.1:5101',
+};